Refactor: update ThreadReactionCard and add cursor style
ThreadReactionCard에 cursor pointer 스타일을 추가 했습니다.
